refactor: replace validation in utils
